refactor: move main.go to cmd/world
