+++ cmd/world/cardinal/purge.go
