+++ cmd/world/cardinal/stop.go
