+++ cmd/world/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"pkg.world.dev/world-cli/cmd/world/root"
+)
+
+func main() {
+	root.Execute()
+}
