--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "pkg.world.dev/world-cli/cmd"
-
-func main() {
-	cmd.Execute()
-}
